pyproject.toml
[project]
name = "data"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.13.1",
    "httpx>=0.28.1",
    "psycopg2-binary>=2.9.9",
    "openai>=1.0.0",
    "sqlalchemy>=2.0.0",
    "python-dotenv>=1.0.0",
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
]
